requirements.txt
tensorflow-gpu==1.10.1
OpenNMT-tf==1.21.4